Put all components into `compoents.rs`. Responding to https://github.com/Inochi2D/inox2d/pull/86#discussion_r1698615893 Rename `ParamMapMode` to `PhysicsParamMapMode` for this change for clarity.
