--- inox2d/src/node/components/composite.rs
@@ -1,4 +1,0 @@
-/// If has this as a component, the node should composite all children
-///
-/// Empty as only a marker, zsorted children list constructed later on demand
-pub struct Composite {}
