--- inox2d/src/node/components/transform_store.rs
@@ -1,10 +1,0 @@
-use glam::Mat4;
-
-use crate::math::transform::TransformOffset;
-
-/// Component holding transform values that may be modified across frames.
-#[derive(Default, Clone)]
-pub struct TransformStore {
-	pub absolute: Mat4,
-	pub relative: TransformOffset,
-}
