+++ inox2d/src/physics/pendulum/rigid.rs
@@ -1,6 +1,6 @@
 use glam::{vec2, Vec2};
 
-use crate::node::components::simple_physics::{PhysicsProps, RigidPendulumCtx};
+use crate::node::components::{PhysicsProps, RigidPendulumCtx};
 use crate::physics::{
 	pendulum::Pendulum,
 	runge_kutta::{IsPhysicsVars, PhysicsState},
